refactor: separate App, Picker, Utils modules and INFO theme; organize WebGuide modules to fields: Pages, Components, Tools
