+++ src/widgets/WebGuide/Components/Status.jsx
